Move combine method to coroutines module
